+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## v8.43.4 (unreleased)
+## v8.43.4 (August 27th 2025)
 ### Implementation changes and bug fixes
+- [PR #2266](https://github.com/rqlite/rqlite/pull/2266): 'go mod' updates.
 - [PR #2220](https://github.com/rqlite/rqlite/pull/2220): Allow `AppendEntriesRequest` to carry application-specific messages.
 - [PR #2221](https://github.com/rqlite/rqlite/pull/2221): Improve the CDC HTTP POST event envelope.
 - [PR #2226](https://github.com/rqlite/rqlite/pull/2226), [PR #2228](https://github.com/rqlite/rqlite/pull/2228): Initial integration between Store and CDC Service.
@@ -17,6 +18,8 @@
 - [PR #2259](https://github.com/rqlite/rqlite/pull/2259): Further improvements to FIFO queue.
 - [PR #2258](https://github.com/rqlite/rqlite/pull/2258): Cluster-level testing of CDC.
 - [PR #2261](https://github.com/rqlite/rqlite/pull/2261): Cluster-level testing of CDC including failure injection.
+- [PR #2262](https://github.com/rqlite/rqlite/pull/2262): Remove redundant persisting HWM to disk.
+- [PR #2265](https://github.com/rqlite/rqlite/pull/2265): rqlite shell now uses Unified Endpoint.  Fixes issue [#257](https://github.com/rqlite/rqlite/issues/257).
 
 ## v8.43.3 (August 14th 2025)
 ### Implementation changes and bug fixes